+++ inst/unitTests/runit.HyperbolicDistribution.R
@@ -57,7 +57,7 @@ function()
     ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
     ## set.seed(4711, kind = "Marsaglia-Multicarry")
     set.seed(4711)
-    test = fBasics:::.distCheck("gh",
+    test = fBasics::distCheck("gh",
         alpha = 1.3, beta = 0.3, delta = 1.7, mu = 0.2, lambda = 0.8,
         n = 2000, robust = FALSE)
     print(test)
@@ -78,7 +78,7 @@ function()
     ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
     ## set.seed(4711, kind = "Marsaglia-Multicarry")
     set.seed(4711)
-    test = fBasics:::.distCheck("hyp",
+    test = fBasics::distCheck("hyp",
         alpha = 1.2, beta = 0.2, delta = 1.9, mu = 0.1, pm = "1",
         n = 1000, robust = FALSE)
     print(test)
@@ -88,7 +88,7 @@ function()
     ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
     ## set.seed(4711, kind = "Marsaglia-Multicarry")
     set.seed(4711)
-    test = fBasics:::.distCheck("hyp",
+    test = fBasics::distCheck("hyp",
         alpha = 0.9, beta = -0.3, delta = 1.4, mu = -0.1, pm = "2",
         n = 1000, robust = FALSE)
     print(test)
@@ -98,7 +98,7 @@ function()
     ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
     ## set.seed(4711, kind = "Marsaglia-Multicarry")
     set.seed(4711)
-    fBasics:::.distCheck("hyp",
+    fBasics::distCheck("hyp",
         alpha = 0.9, beta = -0.3, delta = 1.4, mu = -0.1, pm = "3",
         n = 1000, robust = FALSE)
     print(test)
@@ -108,7 +108,7 @@ function()
     if (FALSE) {
         RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
         set.seed(4711, kind = "Marsaglia-Multicarry")
-        fBasics:::.distCheck("hyp",
+        fBasics::distCheck("hyp",
             alpha = 1.6, beta = -0.3, delta = 1.4, mu = 0.1, pm = "4",
             n = 1000, robust = FALSE)                                    # CHECK
         print(test)
@@ -130,7 +130,7 @@ function()
     ## RNGkind(kind = "Marsaglia-Multicarry", normal.kind = "Inversion")
     ## set.seed(4711, kind = "Marsaglia-Multicarry")
     set.seed(4711)
-    test = fBasics:::.distCheck("nig",
+    test = fBasics::distCheck("nig",
         alpha = 2.1, beta = 0.1, delta = 1.5, mu = -0.1,
         n = 1000, robust = FALSE)
     print(test)