+++ NEWS.md
@@ -1,3 +1,44 @@
+# fBasics 4032.96
+
+- modified `interactivePlot` to allow also functions in argument
+  `plotFun`. Previously `plotFun` had to be a character vector containing the
+  names of the functions. Adjusted the documentation and added an example to
+  this effect.
+
+- function `.distCheck` is defunct, use `distCheck` instead. To allow
+  compatibility with older versions of packages that import fBasics (fGarch,
+  fExtremes), `.distCheck` is exported but new code should not use it.
+
+- removed generic `.print`, its 'ssd' method, and the S3 registration for the
+  latter, all unused.
+
+- clarified in the documentation for `histPlot`, `densityPlot` and
+  `logDensityPlot` that argument `x` must be from class "timeSeries" (each of
+  them starts with `stopifnot(is.timeSeries(x))`). Previously it was stated that
+  argument `x` can be of any class that can be converted to "timeSeries" with
+  `as.timeSeries`.
+
+- contrary to the documentation, the return values from the \code{p/d/q/r*}
+  functions (such as `dhyp`, `phyp`, `qhyp`) for the generalised distributions
+  do not contain attribute \code{"param"}. This is now documented correctly.
+
+- removed attribute 'control' from the return value of `rhyp` and similar
+  functions. It was undocumented (or rather, the documentation stated that it is
+  called "param"; in any case, wouldn't be usable in code).
+
+- argument 'alpha' of `dhyp`, `phyp`, `qhyp` and `rhyp` can now be also a
+  vector of length 4 containing the four parameters.
+
+- `hypMode` now throws error if argument 'pm' doesn't have one of the allowed
+  values.
+
+- the return values of`*Mean()`, `*Var()`, `*Skew()`, `*Kurt()`, `*Moments()`,
+  where '*' is a distribution prefix (such as 'gh' or 'ght'), are now all named.
+  Previously, only some of them were.
+
+- removed a number of inconsistencies in the documentation.
+
+ 
 # fBasics 4031.95
 
 - now `characterTable` doesn't try to print/plot invalid non-ASCII UTF8
@@ -24,7 +65,7 @@
   Willden who suggested the fix.
 
 - the name of the first column of dataset `msft.dat` is not mangled any more (it
-  was "X.Y..m..d" and now is "%Y-%m-%d", as in the input file). the rest of the
+  was "X.Y..m..d" and now is "%Y-%m-%d", as in the input file), the rest of the
   dataset is as before. The file from which the dataset is created is now taken
   from an identical file in package `timeSeries`.
 
@@ -32,8 +73,9 @@
   'Description:' when slot `description = ""`.
 
 - many fitting functions, e.g., `nFit`, were inadvertently ignoring argument
-  `description` by using `description = description()` instead of `description =
-  description` in the call to \code{new} when creating the return value.
+  `description` by using `description = description()` instead of
+  `description = description` in the call to \code{new} when creating the return
+  value.
 
 - `correlationTest`, `pearsonTest`, `spearmanTest`, `kendallTest`, `ks2Test`,
   `scaleTest`, `varianceTest`, and `locationTest` now set slot `description` of